mixed-mode: Add x64 target to mixed-mode assembly tests.  Contributed under MIT/X11 license.
svn path=/trunk/mono/; revision=105847


Commit migrated from https://github.com/mono/mono/commit/9abd67a90706262e6c093ae4415c58dd086f1d8c

+++ src/mono/mono/tests/mixed-mode/MixedModeMsvcrtApp/MixedModeMsvcrtApp.cpp
@@ -16,7 +16,7 @@ void WriteStringManaged (const wchar_t* str)
 void WriteStringUnmanaged (const wchar_t* str)
 {
 	DWORD count;
-	WriteConsole (GetStdHandle (STD_OUTPUT_HANDLE), str, wcslen (str), &count, NULL);
+	WriteConsole (GetStdHandle (STD_OUTPUT_HANDLE), str, (DWORD) wcslen (str), &count, NULL);
 }
 
 void WriteStringWrapper (const wchar_t* str)
